Fixed errror where it wasn't showing messages that didn't have a machine

+++ Factory-Managers/Home-screen.php
@@ -283,19 +283,30 @@
             <div id="inbox-content">
                 <?php
                 require_once "../inc/dbconn.inc.php";
-                $sql = "SELECT n.*, m.name AS machine_name, a.name AS account_name  FROM notes n JOIN machine m ON n.machine_id = m.id JOIN account a ON n.user_id_from = a.id WHERE user_id_to = $_SESSION[userid] ORDER BY note_id DESC";
+                $sql = "SELECT n.*, m.name AS machine_name, a.name AS account_name FROM notes n LEFT JOIN machine m ON n.machine_id = m.id JOIN account a ON n.user_id_from = a.id WHERE user_id_to = $_SESSION[userid] ORDER BY note_id DESC;";
                 $statement = mysqli_stmt_init($conn);
                 mysqli_stmt_prepare($statement, $sql); 
                 if (mysqli_stmt_execute($statement)){
                     $result = mysqli_stmt_get_result($statement);
                     if (mysqli_num_rows($result) >= 1){
                         while ($row = mysqli_fetch_assoc($result)){
-                            echo("
+                            if (isset($row['machine_name'])){
+                                echo("
                                 <details>
                                     <summary>$row[machine_name], $row[notes_subject] <a href='delete-message.php?deletion=$row[note_id]&search-box='>Delete</a></summary>
                                     From: $row[account_name]</br>$row[notes_content]
                                 </details>
                             ");
+                            }
+                            else{
+                                echo("
+                                <details>
+                                    <summary>$row[notes_subject] <a href='delete-message.php?deletion=$row[note_id]&search-box='>Delete</a></summary>
+                                    From: $row[account_name]</br>$row[notes_content]
+                                </details>
+                            ");
+                            }
+                            
                         }
                         mysqli_free_result($result);
                     }
